Move the `Compiler` class into the `Compiler` namespace
