+++ src/Compiler/ObjectCreationCompiler.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace DI\Compiler;
 
-use DI\Compiler;
 use DI\Definition\Exception\InvalidDefinition;
 use DI\Definition\ObjectDefinition;
 use DI\Definition\ObjectDefinition\MethodInjection;
